+++ src/main/java/org/catacombae/dmg/sparseimage/ReadableSparseImageStream.java
@@ -20,7 +20,6 @@ package org.catacombae.dmg.sparseimage;
 import org.catacombae.io.BasicReadableRandomAccessStream;
 import org.catacombae.io.ReadableRandomAccessStream;
 import org.catacombae.io.RuntimeIOException;
-import org.catacombae.io.SynchronizedReadableRandomAccessStream;
 import org.catacombae.util.Util;
 
 /**
@@ -35,9 +34,7 @@ public class ReadableSparseImageStream extends BasicReadableRandomAccessStream {
 
     private long fp = 0;
 
-    public ReadableSparseImageStream(
-            final ReadableRandomAccessStream backingStream)
-    {
+    public ReadableSparseImageStream(final ReadableRandomAccessStream backingStream) {
         this.backingStream = backingStream;
 
         byte[] headerData = new byte[4096];
@@ -46,11 +43,9 @@ public class ReadableSparseImageStream extends BasicReadableRandomAccessStream {
 
         this.header = new SparseImageHeader(headerData, 0);
 
-        final String signature =
-                Util.readString(this.header.getSignature(), "US-ASCII");
-        if(!signature.equals("sprs")) {
-            throw new RuntimeException("Invalid signature: \"" + signature +
-                    "\"");
+        final String signature = Util.readString(this.header.getSignature(), "US-ASCII");
+        if (!signature.equals("sprs")) {
+            throw new RuntimeException("Invalid signature: \"" + signature + "\"");
         }
 
         /* Sector size appears to be fixed at 512. */
@@ -60,36 +55,34 @@ public class ReadableSparseImageStream extends BasicReadableRandomAccessStream {
         this.size = sectorCount * 512;
         this.blockSize = header.getSectorsPerBlock() * 512;
 
-        final long blockMapElementCount =
-                (sectorCount + sectorsPerBlock - 1) / sectorsPerBlock;
-        if(blockMapElementCount > Integer.MAX_VALUE) {
-            throw new RuntimeException("Block map size too large for address " +
-                    "space: " + blockMapElementCount);
+        final long blockMapElementCount = (sectorCount + sectorsPerBlock - 1) / sectorsPerBlock;
+        if (blockMapElementCount > Integer.MAX_VALUE) {
+            throw new RuntimeException("Block map size too large for address " + "space: " + blockMapElementCount);
         }
 
         this.blockMap = new int[(int) blockMapElementCount];
 
-        /* Iterate over the (reverse) block map in header and fill in the
-         * gaps. */
+        /*
+         * Iterate over the (reverse) block map in header and fill in the gaps.
+         */
         int curBlock = 0;
         int curHeader = 0;
-        for(; curBlock < blockMap.length; ++curHeader) {
+        for (; curBlock < blockMap.length; ++curHeader) {
             final int blockMapOffsetInHeader;
             final int blockMapEntriesInHeader;
 
-            if(curHeader == 0) {
+            if (curHeader == 0) {
                 blockMapOffsetInHeader = 64;
                 blockMapEntriesInHeader = 1008;
-            }
-            else {
-                final long nextHeaderOffset =
-                        4096 + 1008 * blockSize +
-                        (curHeader - 1) * (4096 + 1010 * blockSize);
+            } else {
+                final long nextHeaderOffset = 4096 + 1008 * blockSize + (curHeader - 1) * (4096 + 1010 * blockSize);
 
                 backingStream.seek(nextHeaderOffset);
-                if(backingStream.read(headerData) == -1) {
-                    /* If we reach end of file here, all non-hole blocks have
-                     * been processed. The rest are just holes. */
+                if (backingStream.read(headerData) == -1) {
+                    /*
+                     * If we reach end of file here, all non-hole blocks have been processed. The
+                     * rest are just holes.
+                     */
                     break;
                 }
 
@@ -98,31 +91,20 @@ public class ReadableSparseImageStream extends BasicReadableRandomAccessStream {
             }
 
             final int remainingBlocks = blockMap.length - curBlock;
-            final int curEntriesToRead =
-                    remainingBlocks < blockMapEntriesInHeader ?
-                    remainingBlocks : blockMapEntriesInHeader;
+            final int curEntriesToRead = remainingBlocks < blockMapEntriesInHeader ? remainingBlocks : blockMapEntriesInHeader;
 
-            for(int i = 0; i < curEntriesToRead; ++i, ++curBlock) {
-                final long curMapping =
-                        Util.unsign(Util.readIntBE(headerData,
-                        blockMapOffsetInHeader + i * 4));
+            for (int i = 0; i < curEntriesToRead; ++i, ++curBlock) {
+                final long curMapping = Util.unsign(Util.readIntBE(headerData, blockMapOffsetInHeader + i * 4));
 
-                if(curMapping == 0) {
+                if (curMapping == 0) {
                     /* Hole. */
                     continue;
-                }
-                else if((curMapping - 1) > blockMap.length) {
-                    throw new RuntimeException("Inconsistent mapping at " +
-                            curBlock + ": Points at block " + (curMapping - 1) +
-                            ". Number of blocks in sparse image: " +
-                            blockMap.length);
-                }
-                else if(blockMap[(int) curMapping - 1] != 0) {
-                    throw new RuntimeException("Cross-linked mapping: " +
-                            "Physical blocks " +
-                            blockMap[(int) curMapping - 1] + " and " +
-                            curBlock + " both point at virtual block " +
-                            (curMapping - 1) + ".");
+                } else if ((curMapping - 1) > blockMap.length) {
+                    throw new RuntimeException("Inconsistent mapping at " + curBlock + ": Points at block " + (curMapping - 1)
+                            + ". Number of blocks in sparse image: " + blockMap.length);
+                } else if (blockMap[(int) curMapping - 1] != 0) {
+                    throw new RuntimeException("Cross-linked mapping: " + "Physical blocks " + blockMap[(int) curMapping - 1] + " and " + curBlock
+                            + " both point at virtual block " + (curMapping - 1) + ".");
                 }
 
                 blockMap[(int) curMapping - 1] = curBlock + 1;
@@ -137,9 +119,8 @@ public class ReadableSparseImageStream extends BasicReadableRandomAccessStream {
 
     @Override
     public synchronized void seek(final long offset) throws RuntimeIOException {
-        if(offset < 0) {
-            throw new RuntimeIOException("Negative seek offset (" + offset +
-                    ")");
+        if (offset < 0) {
+            throw new RuntimeIOException("Negative seek offset (" + offset + ")");
         }
 
         this.fp = offset;
@@ -156,49 +137,40 @@ public class ReadableSparseImageStream extends BasicReadableRandomAccessStream {
     }
 
     @Override
-    public synchronized int read(final byte[] data, final int pos,
-            final int len) throws RuntimeIOException
-    {
+    public synchronized int read(final byte[] data, final int pos, final int len) throws RuntimeIOException {
         int curPos = pos;
         int remaining;
-        if(fp < size) {
+        if (fp < size) {
             remaining = len > size || fp > size - len ? (int) (size - fp) : len;
-        }
-        else {
+        } else {
             remaining = 0;
         }
 
-        while(remaining > 0) {
+        while (remaining > 0) {
             final long virtualBlockIndex = fp / blockSize;
             final long offsetInBlock = fp % blockSize;
             final long remainingInBlock = blockSize - offsetInBlock;
-            final int bytesToRead =
-                    remaining < remainingInBlock ? remaining :
-                    (int) remainingInBlock;
+            final int bytesToRead = remaining < remainingInBlock ? remaining : (int) remainingInBlock;
 
-            if(virtualBlockIndex >= blockMap.length) {
+            if (virtualBlockIndex >= blockMap.length) {
                 break;
             }
 
             final int blockMapValue = blockMap[(int) virtualBlockIndex];
             int bytesRead = 0;
 
-            if(blockMapValue != 0) {
+            if (blockMapValue != 0) {
                 final int physicalBlockIndex = blockMapValue - 1;
-                final long segmentShift =
-                        4096 + (physicalBlockIndex < 1008 ? 0 :
-                        4096 + ((physicalBlockIndex - 1008) / 1010) * 4096);
-                final long seekOffset = segmentShift +
-                        ((long) physicalBlockIndex) * blockSize +
-                        offsetInBlock;
+                final long segmentShift = 4096 + (physicalBlockIndex < 1008 ? 0 : 4096 + ((physicalBlockIndex - 1008) / 1010) * 4096);
+                final long seekOffset = segmentShift + (physicalBlockIndex) * blockSize + offsetInBlock;
                 backingStream.seek(seekOffset);
                 bytesRead = backingStream.read(data, curPos, bytesToRead);
-                if(bytesRead < 0) {
+                if (bytesRead < 0) {
                     bytesRead = 0;
                 }
             }
 
-            if(bytesRead != bytesToRead) {
+            if (bytesRead != bytesToRead) {
                 /* Not allocated. Just fill with zeroes. */
                 Util.zero(data, curPos + bytesRead, bytesToRead - bytesRead);
                 bytesRead = bytesToRead;
@@ -208,16 +180,15 @@ public class ReadableSparseImageStream extends BasicReadableRandomAccessStream {
             curPos += bytesRead;
             fp += bytesRead;
 
-            if(bytesRead != bytesToRead) {
+            if (bytesRead != bytesToRead) {
                 break;
             }
         }
 
         final int result;
-        if(curPos == pos) {
+        if (curPos == pos) {
             result = -1;
-        }
-        else {
+        } else {
             result = curPos - pos;
         }
